Rename package usecase to interactor for clarity
