+++ src/main/java/interactor/StationState.java
@@ -1,4 +1,4 @@
-package usecase;
+package interactor;
 
 import java.util.List;
